+++ js/data.js
@@ -1,0 +1,28 @@
+const DESCRIPTIONS = [
+  'интересная фотография',
+  'смешная фотокарточка',
+  'забавный перфоманс',
+  'потешная картинка',
+  'классный снимок',
+];
+
+const COMMENTS = [
+  'просто нет слов, когда я зашел на ваш сайт я так сильно обрадовался что просто нет слов',
+  'просто великолепный сайт, я порекомендовал его всем своим друзьям',
+  'сказать что этот сайт просто чудо, это ничего не сказать',
+  'создатели этого сайта, дай Бог вам здоровья, спасибо большое',
+  'после пяти минут на вашем сайте я просто не могу сидеть на других сайтах, невероятный сайт',
+  'как я раньше жил без вашего сайта, теперь моя жизнь стала намного прекрасней',
+];
+
+const NAMES = [
+  'Саша',
+  'Артём',
+  'Никита',
+  'Иван',
+  'Пётр',
+  'Никифор',
+  'Тихон',
+];
+
+export {DESCRIPTIONS, COMMENTS, NAMES};
